requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg[binary]==3.2.*
python-jose[cryptography]>=3.3
passlib>=1.7.4
bcrypt==4.0.1
python-dotenv>=1.0
pydantic>=2.6
pydantic[email]>=2.6